Script: Fix rng

+++ dScripts/newcontent/server/SpawnSkeletonOnDeath.cpp
@@ -5,7 +5,7 @@
 #include "EntityManager.h"
 
 void SpawnSkeletonOnDeath::OnDie(Entity* self, Entity* killer) {
-	auto chance = GeneralUtils::GenerateRandomNumber<uint8_t>(1, 30);
+	auto chance = GeneralUtils::GenerateRandomNumber<uint32_t>(1, 30);
 	if (chance <= 13){
 		EntityInfo info{};
 		info.lot = 20083;
